a user cannot delete another user's thread

+++ forum/app/Http/Controllers/ThreadsController.php
@@ -114,9 +114,10 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
-        //pode ser usado cascade no banco ou
-        //model events ou dessa forma:
-        //$thread->replies()->delete();
+        if($thread->user_id != auth()->id()){
+           abort(403, 'You do not have permission to do this.');
+        }
+
         $thread->delete();
 
         if(request()->wantsJson()) {
